feat: move add and remove association commands from editor.ts to extensions.ts
